jp/requirements.txt
anyio==3.7.1
APScheduler==3.6.3
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
exceptiongroup==1.2.2
fpdf==1.7.2
frozendict==2.4.6
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.3
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.12
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
telegram==0.0.1
tornado==6.4.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.54
